requirements.txt
requests
Flask
python-dotenv
apscheduler
psycopg2-binary

